Refactor providers and their hooks
