+++ client/components/admin/info/BuildEnvironmentSection.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { useTranslation } from '../../../hooks/useTranslation';
+import { useTranslation } from '../../providers/TranslationProvider';
 import { DescriptionList } from './DescriptionList';
 import { formatDate } from './formatters';
