--- client/components/contexts/TranslationContext.js
@@ -1,9 +1,0 @@
-import { createContext } from 'react';
-
-const translate = function(key) {
-	return key;
-};
-
-translate.has = () => true;
-
-export const TranslationContext = createContext(translate);
