+++ client/components/pageNotFound/PageNotFound.js
@@ -1,10 +1,10 @@
 import { Button, ButtonGroup } from '@rocket.chat/fuselage';
 import React from 'react';
 
-import { useTranslation } from '../../hooks/useTranslation';
+import { useTranslation } from '../providers/TranslationProvider';
 import { useWipeInitialPageLoading } from '../../hooks/useWipeInitialPageLoading';
 import { ConnectionStatusAlert } from '../connectionStatus/ConnectionStatusAlert';
-import { useRoute } from '../contexts/RouterContext';
+import { useRoute } from '../providers/RouterProvider';
 import './PageNotFound.css';
 
 export function PageNotFound() {
