+++ client/components/providers/RouterProvider.js
@@ -1,8 +1,15 @@
 import { FlowRouter } from 'meteor/kadira:flow-router';
 import { Tracker } from 'meteor/tracker';
-import React from 'react';
+import React, { createContext, useContext, useEffect, useMemo, useState } from 'react';
 
-import { RouterContext } from '../contexts/RouterContext';
+export const RouterContext = createContext({
+	navigateTo: () => {},
+	replaceWith: () => {},
+	getRouteParameter: () => {},
+	watchRouteParameter: () => {},
+	getQueryStringParameter: () => {},
+	watchQueryStringParameter: () => {},
+});
 
 const navigateTo = (pathDefinition, parameters, queryStringParameters) => {
 	FlowRouter.go(pathDefinition, parameters, queryStringParameters);
@@ -48,3 +55,31 @@ export function RouterProvider({ children }) {
 		{children}
 	</RouterContext.Provider>;
 }
+
+export const useRoute = (pathDefinition) => {
+	const { navigateTo, replaceWith } = useContext(RouterContext);
+
+	return useMemo(() => {
+		const navigate = (...args) => navigateTo(pathDefinition, ...args);
+		navigate.replacingState = (...args) => replaceWith(pathDefinition, ...args);
+		return navigate;
+	}, [navigateTo, replaceWith]);
+};
+
+export const useRouteParameter = (name) => {
+	const { getRouteParameter, watchRouteParameter } = useContext(RouterContext);
+	const [parameter, setParameter] = useState(getRouteParameter(name));
+
+	useEffect(() => watchRouteParameter(name, setParameter), [watchRouteParameter, name]);
+
+	return parameter;
+};
+
+export const useQueryStringParameter = (name) => {
+	const { getQueryStringParameter, watchQueryStringParameter } = useContext(RouterContext);
+	const [parameter, setParameter] = useState(getQueryStringParameter(name));
+
+	useEffect(() => watchQueryStringParameter(name, setParameter), [watchQueryStringParameter, name]);
+
+	return parameter;
+};
