+++ client/components/setupWizard/StateChecker.js
@@ -5,7 +5,7 @@ import { Users } from '../../../app/models';
 import { useSetting } from '../../hooks/useSetting';
 import { useUserId } from '../../hooks/useUserId';
 import { useReactiveValue } from '../../hooks/useReactiveValue';
-import { useRoute } from '../contexts/RouterContext';
+import { useRoute } from '../providers/RouterProvider';
 
 export function StateChecker({ children }) {
 	const setupWizardState = useSetting('Show_Setup_Wizard');
