+++ client/components/setupWizard/StepsState.js
@@ -1,7 +1,7 @@
 import React, { createContext, useContext, useEffect, useMemo, useState } from 'react';
 
 import { useUserId } from '../../hooks/useUserId';
-import { useRouteParameter, useRoute } from '../contexts/RouterContext';
+import { useRouteParameter, useRoute } from '../providers/RouterProvider';
 
 const Context = createContext({});
