+++ client/components/setupWizard/steps/RegisterServerStep.js
@@ -6,8 +6,8 @@ import { handleError } from '../../../../app/utils/client';
 import { useBatchSetSettings } from '../../../hooks/useBatchSetSettings';
 import { useFocus } from '../../../hooks/useFocus';
 import { useMethod } from '../../../hooks/useMethod';
-import { useTranslation } from '../../../hooks/useTranslation';
 import { Icon } from '../../basic/Icon';
+import { useTranslation } from '../../providers/TranslationProvider';
 import { Pager } from '../Pager';
 import { useSetupWizardParameters } from '../ParametersProvider';
 import { Step } from '../Step';
