+++ client/hooks/useEmbeddedLayout.js
@@ -1,3 +1,3 @@
-import { useQueryStringParameter } from '../components/contexts/RouterContext';
+import { useQueryStringParameter } from '../components/providers/RouterProvider';
 
 export const useEmbeddedLayout = () => useQueryStringParameter('layout') === 'embedded';
