--- client/hooks/useTranslation.js
@@ -1,5 +1,0 @@
-import { useContext } from 'react';
-
-import { TranslationContext } from '../components/contexts/TranslationContext';
-
-export const useTranslation = () => useContext(TranslationContext);
